backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
qdrant-client==1.6.9
huggingface_hub==0.12.1
sentence-transformers==2.2.2
openai==1.3.7
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
shopifyapi==12.2.0
numpy==1.24.3
pandas==2.1.3
asyncio==3.4.3
fastapi-cors==0.0.6